+++ gpdb-20160503/alibabacloud_gpdb20160503/models/_list_supabase_projects_response_body.py
@@ -100,6 +100,7 @@ class ListSupabaseProjectsResponseBody(DaraModel):
 class ListSupabaseProjectsResponseBodyItems(DaraModel):
     def __init__(
         self,
+        auto_scale: str = None,
         create_time: str = None,
         dashboard_password: str = None,
         dashboard_user_name: str = None,
@@ -121,6 +122,7 @@ class ListSupabaseProjectsResponseBodyItems(DaraModel):
         vpc_id: str = None,
         zone_id: str = None,
     ):
+        self.auto_scale = auto_scale
         # The creation time.
         self.create_time = create_time
         # Supabase Dashboard password (not used).
@@ -177,6 +179,9 @@ class ListSupabaseProjectsResponseBodyItems(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.auto_scale is not None:
+            result['AutoScale'] = self.auto_scale
+
         if self.create_time is not None:
             result['CreateTime'] = self.create_time
 
@@ -241,6 +246,9 @@ class ListSupabaseProjectsResponseBodyItems(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('AutoScale') is not None:
+            self.auto_scale = m.get('AutoScale')
+
         if m.get('CreateTime') is not None:
             self.create_time = m.get('CreateTime')